support optionally using external ccelerate server for tests

+++ src/tests/integration_tests.cc
@@ -8,9 +8,12 @@
 #include <reproc++/drain.hpp>
 #include <reproc++/reproc.hpp>
 
+#include "../default_endpoint.hh"
+
 namespace ccelerate::tests {
 
 struct Args {
+  bool use_external_server = false;
   std::filesystem::path repo_dir;
   std::filesystem::path test_projects_dir;
 
@@ -30,26 +33,35 @@ static std::string get_random_endpoint() {
 
 class CcelerateServerContext {
 private:
+  bool uses_external_server_;
   reproc::process server_process_;
   std::string endpoint_;
 
 public:
   CcelerateServerContext() {
-    endpoint_ = get_random_endpoint();
-    reproc::options options;
-    const std::vector<std::pair<std::string, std::string>> extra_env = {
-        std::make_pair("CCELERATE_ENDPOINT", endpoint_)};
-    options.env.extra = extra_env;
+    const Args &args = Args::get();
+    uses_external_server_ = args.use_external_server;
+    if (uses_external_server_) {
+      endpoint_ = get_default_ccelerate_endpoint();
+    } else {
+      endpoint_ = get_random_endpoint();
+      reproc::options options;
+      const std::vector<std::pair<std::string, std::string>> extra_env = {
+          std::make_pair("CCELERATE_ENDPOINT", endpoint_)};
+      options.env.extra = extra_env;
 
-    std::vector<std::string> args;
-    args.push_back("./ccelerate");
-    const std::error_code ec = server_process_.start(args, options);
-    EXPECT_FALSE(ec) << "Failed to start server: " << ec.message();
+      std::vector<std::string> args;
+      args.push_back("./ccelerate");
+      const std::error_code ec = server_process_.start(args, options);
+      EXPECT_FALSE(ec) << "Failed to start server: " << ec.message();
+    }
   }
 
   ~CcelerateServerContext() {
-    const std::error_code ec = server_process_.terminate();
-    EXPECT_FALSE(ec) << "Failed to terminate server: " << ec.message();
+    if (!uses_external_server_) {
+      const std::error_code ec = server_process_.terminate();
+      EXPECT_FALSE(ec) << "Failed to terminate server: " << ec.message();
+    }
   }
 
   const std::string &endpoint() const { return endpoint_; }
@@ -98,7 +110,7 @@ TEST(Integration, ClangNoArgs) {
       run_wrapper_process(server_ctx, {"./ccelerate_clang"});
   ASSERT_TRUE(result);
   EXPECT_EQ(result->stdout, "");
-  EXPECT_EQ(result->stderr, "clang: error: no input files\n");
+  EXPECT_NE(result->stderr.find("no input files"), std::string::npos);
   EXPECT_EQ(result->exit_code, 1);
 }
 
@@ -132,6 +144,8 @@ int main(int argc, char **argv) {
   app.add_option("--repo-dir", args.repo_dir,
                  "Path to the ccelerate repository")
       ->required();
+  app.add_flag("--use-external-server", args.use_external_server,
+               "Use an existing external ccelerate server for the tests");
 
   CLI11_PARSE(app, argc, argv);
